Added httapi voicemail application.

+++ voicemail/admin.py
@@ -30,14 +30,16 @@
 from django.contrib import admin
 from django.conf import settings
 from django.forms import ModelForm
-from django.forms.widgets import TextInput
+from django.forms.widgets import TextInput, NumberInput, Select
 from django.contrib.admin.helpers import ACTION_CHECKBOX_NAME
 from import_export.admin import ImportExportModelAdmin
 from import_export import resources
 from pbx.fileabslayer import FileAbsLayer
-from pbx.commonwidgets import PlayerAdminFileFieldWidget
+from pbx.commondestination import CommonDestAction
+from pbx.commonwidgets import PlayerAdminFileFieldWidget, ListTextWidget
 from .models import (
-    Voicemail, VoicemailGreeting,
+    Voicemail, VoicemailGreeting, VoicemailOptions, VoicemailMessages,
+    VoicemailDestinations
 )
 from.voicemailfunctions import VoicemailFunctions
 
@@ -56,13 +58,89 @@ class VoicemailGreetingInlineAdminForm(ModelForm):
 class VoicemailGreetingInLine(admin.TabularInline):
     model = VoicemailGreeting
     form = VoicemailGreetingInlineAdminForm
-
+    classes = ['collapse']
     extra = 1
     fieldsets = [
         (None,          {'fields': ['name', 'filename', 'filestore']}),
     ]
     ordering = [
         'filename'
+    ]
+
+
+class VoicemailMessagesInlineAdminForm(ModelForm):
+
+    class Meta:
+        model = VoicemailMessages
+        widgets = {
+            "name": TextInput(attrs={'size': '40'}),
+            "filename": PlayerAdminFileFieldWidget(),
+            "filestore": TextInput(attrs={'size': '10'}),
+            "caller_id_number": TextInput(attrs={'size': '20'}),
+            "status": TextInput(attrs={'size': '5'}),
+            "duration": NumberInput(attrs={'size': '7'}),
+        }
+        fields = '__all__'
+
+
+class VoicemailMessagesInLine(admin.TabularInline):
+    model = VoicemailMessages
+    form = VoicemailMessagesInlineAdminForm
+
+    extra = 0
+    fieldsets = [
+        (None,          {'fields': ['name', 'filename', 'filestore', 'caller_id_number', 'status', 'read', 'duration']}),
+    ]
+    ordering = [
+        '-created'
+    ]
+
+
+class VoicemailOptionsInlineAdminForm(ModelForm):
+
+    class Meta:
+        model = VoicemailOptions
+        widgets = {
+            "option_param": ListTextWidget(choices=[('', 'List unavailable')], attrs={'size': '50'}),
+        }
+        fields = '__all__'
+
+
+class VoicemailOptionsInLine(admin.TabularInline):
+    model = VoicemailOptions
+    form = VoicemailOptionsInlineAdminForm
+    classes = ['collapse']
+
+    extra = 1
+    fieldsets = [
+        (None,          {'fields': ['option_digits', 'option_param', 'sequence', 'description']}),
+    ]
+    ordering = [
+        'option_digits', 'sequence'
+    ]
+
+
+class VoicemailDestinationsInlineAdminForm(ModelForm):
+
+    class Meta:
+        model = VoicemailDestinations
+        widgets = {
+            "voicemail_dest": Select(choices=[('', 'List unavailable')]),
+        }
+        fields = '__all__'
+
+
+class VoicemailDestinationsInLine(admin.TabularInline):
+    model = VoicemailDestinations
+    form = VoicemailDestinationsInlineAdminForm
+    classes = ['collapse']
+
+    extra = 1
+    fieldsets = [
+        (None,          {'fields': ['voicemail_dest']}),
+    ]
+    ordering = [
+
     ]
 
 
@@ -93,7 +171,7 @@ class VoicemailAdmin(ImportExportModelAdmin):
         # 2. Comment out the def has_add_permission.
         (None,  {'fields': [
                     'password', 'greeting_id', 'alternate_greeting_id',
-                    'mail_to', 'sms_to', 'cc', 'attach_file', 'local_after_email', 'enabled', 'description'
+                    'mail_to', 'sms_to', 'attach_file', 'local_after_email', 'enabled', 'description'
                     ]}),
         ('update Info.',   {'fields': ['created', 'updated', 'synchronised', 'updated_by'], 'classes': ['collapse']}),
     ]
@@ -104,19 +182,34 @@ class VoicemailAdmin(ImportExportModelAdmin):
         'extension_id__extension'
     ]
     actions = [sync_greeting_files]
-    inlines = [VoicemailGreetingInLine]
+    inlines = [VoicemailGreetingInLine, VoicemailOptionsInLine, VoicemailDestinationsInLine, VoicemailMessagesInLine]
 
     def has_add_permission(self, request):
         return False
+
+    def get_formsets_with_inlines(self, request, obj=None):
+        vma = CommonDestAction(request.session['domain_name'], request.session['domain_uuid'])
+        option_param_choices = vma.get_action_choices(' ', True, 0b1011111111111111)
+        vm_destination_choices = [('', '------')]
+        qs = Voicemail.objects.filter(enabled='true', extension_id__domain_id_id=request.session['domain_uuid']).order_by('extension_id')
+        for q in qs:
+            vm_destination_choices.append((str(q.id), q.extension_id))
+        for inline in self.get_inline_instances(request, obj):
+            if type(inline) is VoicemailOptionsInLine:
+                inline.form.Meta.widgets['option_param'].choices=option_param_choices
+            if type(inline) is VoicemailDestinationsInLine:
+                inline.form.Meta.widgets['voicemail_dest'].choices=vm_destination_choices
+            yield inline.get_formset(request, obj), inline
 
     def save_formset(self, request, form, formset, change):
         if not settings.PBX_FREESWITCH_LOCAL:
             fal = FileAbsLayer()
         instances = formset.save(commit=False)
         for obj in formset.deleted_objects:
-            if not settings.PBX_FREESWITCH_LOCAL:
-                fal.delete('/home/django-pbx/media/%s' % obj.filename.name, request.session['home_switch'])
-            obj.filename.delete(save=False)
+            if type(obj) is VoicemailGreeting or type(obj) is VoicemailMessages:
+                if not settings.PBX_FREESWITCH_LOCAL:
+                    fal.delete('/home/django-pbx/media/%s' % obj.filename.name, request.session['home_switch'])
+                obj.filename.delete(save=False)
             obj.delete()
         for instance in instances:
             instance.updated_by = request.user.username
